fix(eip-7702): filter_invalid_txs intrinsic gas check (gravity-audit#668)
The pre-execution filter now mirrors the pool's ensure_intrinsic_gas
so a TxEip7702 with gas_limit < 21000 + 25000·N — injected outside the
pool's fork-gate — is discarded instead of reaching grevm and panicking
the executor with IntrinsicGasTooLow.

Free filter_invalid_txs and its tests move into a new tx_filter module
per _local/drafts/pipe-exec-layer-lib-refactor.md step 3, shrinking
lib.rs from 1829 to 1448 lines. Two regression tests added for
intrinsic-gas-too-low and at-the-floor 7702 cases under Prague.
