+++ crates/pipe-exec-layer-ext-v2/execute/Cargo.toml
@@ -27,6 +27,7 @@ reth-trie-parallel.workspace = true
 reth-provider.workspace = true
 reth-rpc-eth-api.workspace = true
 revm-primitives.workspace = true
+revm-interpreter.workspace = true
 gravity-storage.workspace = true
 gravity-primitives.workspace = true
 grevm.workspace = true
